+++ src/android/com/appsflyer/cordova/plugin/AppsFlyerConstants.java
@@ -14,6 +14,10 @@ public class AppsFlyerConstants {
 
 
     final static String AF_IS_DEBUG = "isDebug";
+
+    final static String AF_COLLECT_IMEI = "collectIMEI";
+    final static String AF_COLLECT_ANDROID_ID = "collectAndroidID";
+
     final static String AF_DEV_KEY = "devKey";
     final static String AF_CONVERSION_DATA = "onInstallConversionDataListener";
     final static String AF_ON_INSTALL_CONVERSION_DATA = "onInstallConversionDataN";
@@ -24,4 +28,14 @@ public class AppsFlyerConstants {
     final static String AF_ON_APP_OPEN_ATTRIBUTION ="onAppOpenAttribution";
     final static String AF_ON_INSTALL_CONVERSION_FAILURE ="onInstallConversionFailure";
     final static String AF_ON_INSTALL_CONVERSION_DATA_LOADED ="onInstallConversionDataLoaded";
+
+    
+    final static String INVITE_FAIL = "Could not create invite link";
+    final static String INVITE_CHANNEL = "channel";
+    final static String INVITE_CAMPAIGN = "campaign";
+    final static String INVITE_REFERRER = "referrerName";
+    final static String INVITE_IMAGEURL = "referreImageURL";
+    final static String INVITE_CUSTOMERID = "customerID";
+    final static String INVITE_DEEPLINK = "baseDeepLink";
+    final static String PROMOTE_ID = "promotedAppId";
 }