+++ src/android/com/appsflyer/cordova/plugin/AppsFlyerPlugin.java
@@ -1,5 +1,6 @@
 package com.appsflyer.cordova.plugin;
 
+import android.net.Uri;
 import java.util.HashMap;
 import java.util.Iterator;
 import java.util.Map;
@@ -16,22 +17,33 @@ import org.json.JSONException;
 import com.appsflyer.AppsFlyerConversionListener;
 import com.appsflyer.AppsFlyerLib;
 import com.appsflyer.AppsFlyerProperties;
+import com.appsflyer.CreateOneLinkHttpTask;
+import com.appsflyer.share.CrossPromotionHelper;
+import com.appsflyer.share.LinkGenerator;
+import com.appsflyer.share.ShareInviteHelper;
 
+import android.app.Activity;
 import android.content.Context;
 import android.content.Intent;
 import android.util.Log;
-import android.os.Build;
+
 
 import static com.appsflyer.cordova.plugin.AppsFlyerConstants.*;
 
 public class AppsFlyerPlugin extends CordovaPlugin {
 
 	private CallbackContext mConversionListener = null;
+    private CallbackContext mAttributionDataListener = null;
+    private Map<String, String> mAttributionData = null;
+    private CallbackContext mInviteListener = null;
+    private Uri intentURI = null;
+    private Uri newIntentURI = null;
+    private Activity c;
 
 	@Override
 	public void initialize(CordovaInterface cordova, CordovaWebView webView) {
 		super.initialize(cordova, webView);
-	}
+    }
 
 	/**
 	 * Called when the activity receives a new intent.
@@ -39,7 +51,7 @@ public class AppsFlyerPlugin extends CordovaPlugin {
 	@Override
 	public void onNewIntent(Intent intent) {
 		cordova.getActivity().setIntent(intent);
-		AppsFlyerLib.getInstance().sendDeepLinkData(cordova.getActivity());
+        AppsFlyerLib.getInstance().sendDeepLinkData(cordova.getActivity());
 	}
 
 	@Override
@@ -49,6 +61,9 @@ public class AppsFlyerPlugin extends CordovaPlugin {
 		{
 			return setCurrencyCode(args);
 		}
+		else if("registerOnAppOpenAttribution".equals(action)){
+			return registerOnAppOpenAttribution(callbackContext);
+		}
 		else if("setAppUserId".equals(action))
 		{
 			return setAppUserId(args, callbackContext);
@@ -56,6 +71,12 @@ public class AppsFlyerPlugin extends CordovaPlugin {
 		else if("getAppsFlyerUID".equals(action))
 		{
 			return getAppsFlyerUID(callbackContext);
+		}
+		else if ("setDeviceTrackingDisabled".equals(action)) {
+			return setDeviceTrackingDisabled(args);
+		}
+		else if ("stopTracking".equals(action)) {
+		    return stopTracking(args);
 		}
 		else if("initSdk".equals(action))
 		{
@@ -71,16 +92,43 @@ public class AppsFlyerPlugin extends CordovaPlugin {
 		{
 			return enableUninstallTracking(args, callbackContext);
 		}
+		else if ("updateServerUninstallToken".equals(action)) {
+			return updateServerUninstallToken(args, callbackContext);
+		}
+		else if ("setAppInviteOneLinkID".equals(action)) {
+			return setAppInviteOneLinkID(args, callbackContext);
+		}
+		else if ("generateInviteLink".equals(action)) {
+            return generateInviteLink(args, callbackContext);
+        }
+        else if ("trackCrossPromotionImpression".equals(action)) {
+        	return trackCrossPromotionImpression(args, callbackContext);
+        }
+        else if ("trackAndOpenStore".equals(action)) {
+            return trackAndOpenStore(args, callbackContext);
+        }
+		else if("resumeSDK".equals(action))
+		{
+			return onResume(args, callbackContext);
+		}
 
 		return false;
 	}
 
-
-
-    private void trackAppLaunch(){
-        Context c = this.cordova.getActivity().getApplicationContext();
+    private void trackAppLaunch() {
+        c = this.cordova.getActivity();
+        AppsFlyerLib.getInstance().sendDeepLinkData(c);
         AppsFlyerLib.getInstance().trackEvent(c, null, null);
     }
+
+    private boolean registerOnAppOpenAttribution(final CallbackContext callbackContext){
+
+		if(mAttributionDataListener == null) {
+			mAttributionDataListener = callbackContext;
+		}
+
+		return true;
+	}
 
 	/**
 	 *
@@ -93,6 +141,7 @@ public class AppsFlyerPlugin extends CordovaPlugin {
 		String devKey = null;
 		boolean isConversionData;
 		boolean isDebug = false;
+		AppsFlyerConversionListener gcdListener = null;
 
 		AppsFlyerProperties.getInstance().set(AppsFlyerProperties.LAUNCH_PROTECT_ENABLED, false);
 		AppsFlyerLib instance = AppsFlyerLib.getInstance();
@@ -109,27 +158,48 @@ public class AppsFlyerPlugin extends CordovaPlugin {
 
 			isDebug = options.optBoolean(AF_IS_DEBUG, false);
 
+			if(options.has(AF_COLLECT_ANDROID_ID)){
+				AppsFlyerLib.getInstance().setCollectAndroidID(options.optBoolean(AF_COLLECT_ANDROID_ID, true));
+			}
+			if(options.has(AF_COLLECT_IMEI)){
+				AppsFlyerLib.getInstance().setCollectIMEI(options.optBoolean(AF_COLLECT_IMEI, true));
+			}
+
+
+
 			instance.setDebugLog(isDebug);
 
-			if(isDebug == true){ Log.d("AppsFlyer", "Starting Tracking");}
-			trackAppLaunch();
-
-			instance.startTracking(AppsFlyerPlugin.this.cordova.getActivity().getApplication(), devKey);
-
+			if(isDebug == true){
+                Log.d("AppsFlyer", "Starting Tracking");
+			}
 
 			if(isConversionData == true){
+
+//				if(mAttributionDataListener == null) {
+//					mAttributionDataListener = callbackContext;
+//				}
 
 				if(mConversionListener == null){
 					mConversionListener = callbackContext;
 				}
 
-				registerConversionListener(instance);
-				sendPluginNoResult(callbackContext);
+				gcdListener = registerConversionListener(instance);
 			}
 			else{
 				callbackContext.success(SUCCESS);
 			}
 
+			instance.init(devKey, gcdListener, cordova.getActivity());
+
+			trackAppLaunch();
+			instance.startTracking(c.getApplication());
+
+			if(gcdListener != null){
+				sendPluginNoResult(callbackContext);
+			}
+			else{
+				callbackContext.success(SUCCESS);
+			}
 		}
 		catch (JSONException e){
 			e.printStackTrace();
@@ -138,12 +208,18 @@ public class AppsFlyerPlugin extends CordovaPlugin {
 		return true;
 	}
 
-	private void registerConversionListener(AppsFlyerLib instance){
-		instance.registerConversionListener(cordova.getActivity().getApplicationContext(), new AppsFlyerConversionListener(){
+	private AppsFlyerConversionListener registerConversionListener(AppsFlyerLib instance){
+		return new AppsFlyerConversionListener(){
 
 			@Override
 			public void onAppOpenAttribution(Map<String, String> attributionData) {
-				handleSuccess(AF_ON_APP_OPEN_ATTRIBUTION, attributionData);
+                mAttributionData = attributionData;
+                intentURI =  c.getIntent().getData();
+
+				handleSuccess(AF_ON_APP_OPEN_ATTRIBUTION, mAttributionData);
+
+
+
 			}
 
 			@Override
@@ -154,6 +230,7 @@ public class AppsFlyerPlugin extends CordovaPlugin {
 			@Override
 			public void onInstallConversionDataLoaded(Map<String, String> conversionData) {
 				handleSuccess(AF_ON_INSTALL_CONVERSION_DATA_LOADED, conversionData);
+
 			}
 
 			@Override
@@ -195,8 +272,22 @@ public class AppsFlyerPlugin extends CordovaPlugin {
 
 				final String jsonStr = params.toString();
 
+				if(
+						( params.optString("type") == AF_ON_ATTRIBUTION_FAILURE
+						||params.optString("type") == AF_ON_APP_OPEN_ATTRIBUTION )
+								&& mAttributionDataListener != null)
+				{
+					PluginResult result = new PluginResult(PluginResult.Status.OK, jsonStr);
+					result.setKeepCallback(false);
 
-				if (mConversionListener != null) {
+					mAttributionDataListener.sendPluginResult(result);
+					mAttributionDataListener = null;
+				}
+				else if(
+						(params.optString("type") == AF_ON_INSTALL_CONVERSION_DATA_LOADED
+						|| params.optString("type") == AF_ON_INSTALL_CONVERSION_FAILURE)
+								&&  mConversionListener != null)
+				{
 					PluginResult result = new PluginResult(PluginResult.Status.OK, jsonStr);
 					result.setKeepCallback(false);
 
@@ -204,7 +295,7 @@ public class AppsFlyerPlugin extends CordovaPlugin {
 					mConversionListener = null;
 				}
 			}
-		});
+		};
 	}
 
 	private boolean trackEvent(JSONArray parameters, final CallbackContext callbackContext) {
@@ -287,6 +378,36 @@ public class AppsFlyerPlugin extends CordovaPlugin {
 		return true;
 	}
 
+	private boolean setDeviceTrackingDisabled(JSONArray parameters){
+
+		try
+		{
+			boolean isDisabled = parameters.getBoolean(0);
+			AppsFlyerLib.getInstance().setDeviceTrackingDisabled(isDisabled);
+		}
+		catch (JSONException e)
+		{
+			e.printStackTrace();
+			return true; //TODO error
+		}
+		return true;
+	}
+
+		private boolean stopTracking(JSONArray parameters){
+
+    		try
+    		{
+    			boolean isStopTracking = parameters.getBoolean(0);
+    			AppsFlyerLib.getInstance().stopTracking(isStopTracking, cordova.getActivity().getApplicationContext());
+    		}
+    		catch (JSONException e)
+    		{
+    			e.printStackTrace();
+    			return true; //TODO error
+    		}
+    		return true;
+    	}
+
 	private static Map<String,Object> jsonToMap(String inputString){
 		Map<String,Object> newMap = new HashMap<String, Object>();
 
@@ -322,6 +443,20 @@ public class AppsFlyerPlugin extends CordovaPlugin {
 		return true;
 	}
 
+	private boolean updateServerUninstallToken(JSONArray parameters, CallbackContext callbackContext) {
+		String token = parameters.optString(0);
+		if (token != null && token.length() > 0) {
+			Context c = this.cordova.getActivity().getApplicationContext();
+			AppsFlyerLib.getInstance().updateServerUninstallToken(c,token);
+			callbackContext.success(SUCCESS);
+			return true;
+		}
+		else {
+			callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.ERROR, "Not a valid token"));
+		}
+		return true;
+	}
+
 	private boolean enableUninstallTracking(JSONArray parameters, CallbackContext callbackContext){
 
 		String gcmProjectNumber = parameters.optString(0);
@@ -335,6 +470,177 @@ public class AppsFlyerPlugin extends CordovaPlugin {
 		callbackContext.success(SUCCESS);
 		return true;
 	}
+
+    private boolean onResume(JSONArray parameters, CallbackContext callbackContext){
+        Intent intent = cordova.getActivity().getIntent();
+        newIntentURI = intent.getData();
+
+        if (newIntentURI != intentURI) {
+            if (mAttributionData != null) {
+                PluginResult r = new PluginResult(PluginResult.Status.OK, new JSONObject(mAttributionData).toString());
+                callbackContext.sendPluginResult(r);
+                mAttributionData = null;
+            } else {
+                mAttributionDataListener = callbackContext;
+                sendPluginNoResult(callbackContext);
+            }
+            
+            intentURI = newIntentURI;
+        }
+        return true;
+    }
+
+    // USER INVITE
+	private boolean setAppInviteOneLinkID(JSONArray parameters, CallbackContext callbackContext) {
+		try {
+			String oneLinkID = parameters.getString(0);
+			if (oneLinkID == null || oneLinkID.length() == 0) {
+				return true; //TODO error
+			}
+			AppsFlyerLib.getInstance().setAppInviteOneLink(oneLinkID);
+			callbackContext.success(SUCCESS);
+		} catch (JSONException e) {
+			e.printStackTrace();
+			return true; //TODO error
+		}
+		return true;
+	}
+
+	private boolean generateInviteLink(JSONArray args, CallbackContext callbackContext) {
+
+        String channel = null;
+        String campaign = null;
+        String referrerName = null;
+        String referrerImageUrl = null;
+        String customerID = null;
+        String baseDeepLink = null;
+
+        try {
+            final JSONObject options = args.getJSONObject(0);
+
+            channel = options.optString(INVITE_CHANNEL, "");
+            campaign = options.optString(INVITE_CAMPAIGN, "");
+            referrerName = options.optString(INVITE_REFERRER, "");
+            referrerImageUrl = options.optString(INVITE_IMAGEURL, "");
+            customerID = options.optString(INVITE_CUSTOMERID, "");
+            baseDeepLink = options.optString(INVITE_DEEPLINK, "");
+
+			Context context = this.cordova.getActivity().getApplicationContext();
+			LinkGenerator linkGenerator = ShareInviteHelper.generateInviteUrl(context);
+
+            if (channel !=null && channel != ""){
+                linkGenerator.setChannel(channel);
+            }
+            if (campaign !=null && campaign != ""){
+                linkGenerator.setCampaign(campaign);
+            }
+            if (referrerName !=null && referrerName != ""){
+                linkGenerator.setReferrerName(referrerName);
+            }
+            if (referrerImageUrl !=null && referrerImageUrl != ""){
+                linkGenerator.setReferrerImageURL(referrerImageUrl);
+            }
+            if (customerID !=null && customerID != ""){
+                linkGenerator.setReferrerCustomerId(customerID);
+            }
+            if (baseDeepLink !=null && baseDeepLink != ""){
+                linkGenerator.setBaseDeeplink(baseDeepLink);
+            }
+
+            if (options.length() > 1 && !options.get("userParams").equals("")) {
+                JSONObject jsonCustomValues = options.getJSONObject("userParams");
+
+                Iterator<?> keys = jsonCustomValues.keys();
+
+                while( keys.hasNext() ) {
+                    String key = (String)keys.next();
+                    Object keyvalue = jsonCustomValues.get(key);
+                    linkGenerator.addParameter(key, keyvalue.toString());
+                }
+            }
+
+			linkGenerator.generateLink(context, new inviteCallbacksImpl());
+            mInviteListener = callbackContext;
+            sendPluginNoResult(mInviteListener);
+		}
+		catch (JSONException e) {
+            callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.ERROR, INVITE_FAIL));
+		}
+        return  true;
+    }
+
+    private class inviteCallbacksImpl implements CreateOneLinkHttpTask.ResponseListener {
+        @Override
+        public void onResponse(String s) {
+            PluginResult result = new PluginResult(PluginResult.Status.OK, s);
+            result.setKeepCallback(false);
+            mInviteListener.sendPluginResult(result);
+        }
+
+        @Override
+        public void onResponseError(String s) {
+
+        }
+    }
+
+    // CROSS PROMOTION
+    public boolean trackCrossPromotionImpression(JSONArray parameters, CallbackContext callbackContext) {
+    	String promotedAppId = null;
+        String campaign = null;
+
+        try {
+            final JSONObject options = parameters.getJSONObject(0);
+
+            promotedAppId = options.optString(PROMOTE_ID, "");
+            campaign = options.optString(INVITE_CAMPAIGN, "");
+
+            if (promotedAppId !=null && promotedAppId != "") {
+                Context context = this.cordova.getActivity().getApplicationContext();
+                CrossPromotionHelper.trackCrossPromoteImpression(context,promotedAppId,campaign);
+            }
+            else {
+                callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.ERROR, "CrossPromoted App ID Not set"));
+                return true;
+            }
+
+        } catch (JSONException e) {
+            callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.ERROR, "CrossPromotionImpression Failed"));
+        }
+        return true;
+
+    }
+
+           public boolean trackAndOpenStore(JSONArray parameters, CallbackContext callbackContext) {
+        String promotedAppId = null;
+        String campaign = null;
+        Map<String, String> userParams = null;
+        try {
+            promotedAppId = parameters.getString(0);
+            campaign = parameters.getString(1);
+            if (promotedAppId !=null && promotedAppId != "") {
+                Context context = this.cordova.getActivity().getApplicationContext();
+                if (!parameters.isNull(2)) {
+                    Map<String,String> newUserParams = new HashMap<String,String>();
+                    JSONObject usrParams = parameters.optJSONObject(2);
+                    Iterator<?> keys = usrParams.keys();
+                    while (keys.hasNext()) {
+                        String key = (String) keys.next();
+                        Object keyvalue = usrParams.get(key);
+                        newUserParams.put(key,keyvalue.toString());
+                    }
+                    userParams = newUserParams;
+                }
+                CrossPromotionHelper.trackAndOpenStore(context,promotedAppId,campaign, userParams);
+            }
+            else {
+                callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.ERROR, "CrossPromoted App ID Not set"));
+                return true;
+            }
+        } catch (JSONException e) {
+            callbackContext.sendPluginResult(new PluginResult(PluginResult.Status.ERROR, "CrossPromotion Failed"));
+        }
+        return true;
+    }
 
 	private void sendPluginNoResult(CallbackContext callbackContext) {
 		PluginResult pluginResult = new PluginResult(