+++ src/ios/AppsFlyerPlugin.h
@@ -4,6 +4,8 @@
 
 
 @interface AppsFlyerPlugin : CDVPlugin <UIApplicationDelegate, AppsFlyerTrackerDelegate>
+- (void)initSdk:(CDVInvokedUrlCommand*)command;
+- (void)resumeSDK:(CDVInvokedUrlCommand *)command;
 - (void)setCurrencyCode:(CDVInvokedUrlCommand*)command;
 - (void)setAppUserId:(CDVInvokedUrlCommand*)command;
 - (void)getAppsFlyerUID:(CDVInvokedUrlCommand*)command;
@@ -13,6 +15,13 @@
 - (void)trackEvent:(CDVInvokedUrlCommand*)command;
 - (void)registerUninstall:(CDVInvokedUrlCommand*)command;
 - (void)handleOpenUrl:(CDVInvokedUrlCommand *)url;
+- (void)setDeviceTrackingDisabled:(CDVInvokedUrlCommand *)command;
+- (void)stopTracking:(CDVInvokedUrlCommand *) command;
+- (void)setAppInviteOneLinkID:(CDVInvokedUrlCommand *)command;
+- (void)generateInviteLink:(CDVInvokedUrlCommand*)command;
+- (void)trackCrossPromotionImpression:(CDVInvokedUrlCommand *)command;
+- (void)trackAndOpenStore:(CDVInvokedUrlCommand *)command;
+- (void)registerOnAppOpenAttribution:(CDVInvokedUrlCommand *)command;
 @end
 
 
@@ -21,7 +30,16 @@
 // Appsflyer JS objects
 #define afDevKey                        @"devKey"
 #define afAppId                         @"appId"
-#define afIsDebug                       @"isDebug"
+#define afIsDebug						@"isDebug"
+
+// User Invites, Cross Promotion
+#define afCpAppID                       @"crossPromotedAppId"
+#define afUiChannel                     @"channel"
+#define afUiCampaign                    @"campaign"
+#define afUiRefName                     @"referrerName"
+#define afUiImageUrl                    @"referrerImageUrl"
+#define afUiCustomerID                  @"customerID"
+#define afUiBaseDeepLink                @"baseDeepLink"
 
 // Appsflyer native objects
 #define afConversionData                @"onInstallConversionDataListener"