+++ src/ios/AppsFlyerPlugin.m
@@ -1,5 +1,6 @@
 #import "AppsFlyerPlugin.h"
 #import "AppsFlyerTracker.h"
+#import "AppDelegate.h"
 
 @implementation AppsFlyerPlugin
 
@@ -9,7 +10,11 @@ static NSString *const NO_APPID_FOUND  = @"'appId' is missing or empty";
 static NSString *const SUCCESS         = @"Success";
 
  NSString* mConversionListener;
-
+ NSString* mAttributionDataListener;
+ NSString* mConversionListenerOnResume;
+ NSString* mInviteListener;
+ BOOL isConversionData = NO;
+    
 - (void)pluginInitialize{}
 
 - (void)initSdk:(CDVInvokedUrlCommand*)command
@@ -19,7 +24,7 @@ static NSString *const SUCCESS         = @"Success";
     NSString* devKey = nil;
     NSString* appId = nil;
     BOOL isDebug = NO;
-    BOOL isConversionData = NO;
+    
     
     if (![initSdkOptions isKindOfClass:[NSNull class]]) {
         
@@ -75,7 +80,26 @@ static NSString *const SUCCESS         = @"Success";
         }
     }
   }
+    
+- (void)resumeSDK:(CDVInvokedUrlCommand *)command
+  {
+      [[AppsFlyerTracker sharedTracker] trackAppLaunch];
+      
+      
+      if (isConversionData == YES) {
+          CDVPluginResult* pluginResult = nil;
+          mConversionListenerOnResume = command.callbackId;
+          
+          pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+          [pluginResult setKeepCallback:[NSNumber numberWithBool:YES]];
+      }
+      else {
+          CDVPluginResult* result = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsString:SUCCESS];
+          [self.commandDelegate sendPluginResult:result callbackId:command.callbackId];
+      }
+  }
 
+    
 - (void)setCurrencyCode:(CDVInvokedUrlCommand*)command
 {
     if ([command.arguments count] == 0) {
@@ -96,6 +120,36 @@ static NSString *const SUCCESS         = @"Success";
     [AppsFlyerTracker sharedTracker].customerUserID  = userId;
 }
 
+- (void)setDeviceTrackingDisabled:(CDVInvokedUrlCommand *)command
+{
+    if ([command.arguments count] == 0) {
+        return;
+    }
+    
+    BOOL isDisValueBool = NO;
+    id isDisValue = nil;
+    isDisValue = [command.arguments objectAtIndex:0];
+    if ([isDisValue isKindOfClass:[NSNumber class]]) {
+        isDisValueBool = [(NSNumber*)isDisValue boolValue];
+        [AppsFlyerTracker sharedTracker].deviceTrackingDisabled  = isDisValueBool;
+    }
+}
+
+- (void)stopTracking:(CDVInvokedUrlCommand *)command
+{
+    if ([command.arguments count] == 0) {
+        return;
+    }
+
+    BOOL isStopValueBool = NO;
+    id isStopValue = nil;
+    isStopValue = [command.arguments objectAtIndex:0];
+    if ([isStopValue isKindOfClass:[NSNumber class]]) {
+        isStopValueBool = [(NSNumber*)isStopValue boolValue];
+        [AppsFlyerTracker sharedTracker].isStopTracking  = isStopValueBool;
+    }
+}
+
 - (void)getAppsFlyerUID:(CDVInvokedUrlCommand *)command
 {
     NSString* userId = [[AppsFlyerTracker sharedTracker] getAppsFlyerUID];
@@ -104,6 +158,13 @@ static NSString *const SUCCESS         = @"Success";
                                     messageAsString: userId
                                     ];
     [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+}
+
+- (void)registerOnAppOpenAttribution:(CDVInvokedUrlCommand *)command
+{
+    mAttributionDataListener = command.callbackId;
+    CDVPluginResult *pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+    [pluginResult setKeepCallback:[NSNumber numberWithBool:YES]];
 }
 
 - (void)sendTrackingWithEvent:(CDVInvokedUrlCommand *)command
@@ -138,6 +199,119 @@ static NSString *const SUCCESS         = @"Success";
     }
 }
 
+//USER INVITES
+    
+- (void)setAppInviteOneLinkID:(CDVInvokedUrlCommand*)command {
+    if ([command.arguments count] == 0) {
+        return;
+    }
+    NSString* oneLinkID = [command.arguments objectAtIndex:0];
+    [AppsFlyerTracker sharedTracker].appInviteOneLinkID = oneLinkID;
+}
+    
+- (void)generateInviteLink:(CDVInvokedUrlCommand*)command {
+    NSDictionary* inviteLinkOptions = [command argumentAtIndex:0 withDefault:[NSNull null]];
+    NSDictionary* customParams = [command argumentAtIndex:1 withDefault:[NSNull null]];
+    
+    NSString *channel = nil;
+    NSString *campaign = nil;
+    NSString *referrerName = nil;
+    NSString *referrerImageUrl = nil;
+    NSString *customerID = nil;
+    NSString *baseDeepLink = nil;
+    
+    if (![inviteLinkOptions isKindOfClass:[NSNull class]]) {
+        channel = (NSString*)[inviteLinkOptions objectForKey: afUiChannel];
+        campaign = (NSString*)[inviteLinkOptions objectForKey: afUiCampaign];
+        referrerName = (NSString*)[inviteLinkOptions objectForKey: afUiRefName];
+        referrerImageUrl = (NSString*)[inviteLinkOptions objectForKey: afUiImageUrl];
+        customerID = (NSString*)[inviteLinkOptions objectForKey: afUiCustomerID];
+        baseDeepLink = (NSString*)[inviteLinkOptions objectForKey: afUiBaseDeepLink];
+        
+        [AppsFlyerShareInviteHelper generateInviteUrlWithLinkGenerator:^AppsFlyerLinkGenerator * _Nonnull(AppsFlyerLinkGenerator * _Nonnull generator) {
+            if (channel != nil && ![channel isEqualToString:@""]) {
+                [generator setChannel:channel];
+            }
+            if (campaign != nil && ![campaign isEqualToString:@""]) {
+                [generator setCampaign:campaign];
+            }
+            if (referrerName != nil && ![referrerName isEqualToString:@""]) {
+                [generator setReferrerName:referrerName];
+            }
+            if (referrerImageUrl != nil && ![referrerImageUrl isEqualToString:@""]) {
+                [generator setReferrerImageURL:referrerImageUrl];
+            }
+            if (customerID != nil && ![customerID isEqualToString:@""]) {
+                [generator setReferrerCustomerId:customerID];
+            }
+            if (baseDeepLink != nil && ![baseDeepLink isEqualToString:@""]) {
+                [generator setDeeplinkPath:baseDeepLink];
+            }
+            
+            if (![customParams isKindOfClass:[NSNull class]]) {
+                    [generator addParameters:customParams];
+            }
+            
+            return generator;
+        } completionHandler: ^(NSURL * _Nullable url) {
+            mInviteListener = url.absoluteString;
+                if (mInviteListener != nil) {
+                CDVPluginResult *pluginResult = [ CDVPluginResult
+                                                 resultWithStatus    : CDVCommandStatus_OK
+                                                 messageAsString: mInviteListener
+                                                 ];
+                
+                [self.commandDelegate sendPluginResult:pluginResult callbackId:command.callbackId];
+            }
+         }];
+    }
+}
+    
+//CROSS PROMOTION
+-(void)trackCrossPromotionImpression:(CDVInvokedUrlCommand*) command {
+    
+    if ([command.arguments count] == 0) {
+        return;
+    }
+    
+    NSString* campaign = nil;
+    NSString* promtAppID = [command.arguments objectAtIndex:0];
+    campaign = [command.arguments objectAtIndex:1];
+    
+    if (promtAppID != nil && ![promtAppID isEqualToString:@""]) {
+        [AppsFlyerCrossPromotionHelper trackCrossPromoteImpression:promtAppID campaign:campaign];
+    }
+}
+
+-(void)trackAndOpenStore:(CDVInvokedUrlCommand*) command {
+    
+    if ([command.arguments count] == 0) {
+        return;
+    }
+
+    NSString* promtAppID = [command.arguments objectAtIndex:0];
+    NSString* campaign = [command.arguments objectAtIndex:1];
+    NSDictionary* customParams = [command argumentAtIndex:2 withDefault:[NSNull null]];
+    
+    if (promtAppID != nil && ![promtAppID isEqualToString:@""]) {
+        [AppsFlyerShareInviteHelper generateInviteUrlWithLinkGenerator:^AppsFlyerLinkGenerator * _Nonnull(AppsFlyerLinkGenerator * _Nonnull generator) {
+            if (campaign != nil && ![campaign isEqualToString:@""]) {
+                [generator setCampaign:campaign];
+            }
+            if (![customParams isKindOfClass:[NSNull class]]) {
+                [generator addParameters:customParams];
+            }
+            
+            return generator;
+        } completionHandler: ^(NSURL * _Nullable url) {
+            NSString *appLink = url.absoluteString;
+            [[UIApplication sharedApplication] openURL:[NSURL URLWithString:appLink] options:@{} completionHandler:^(BOOL success) {
+                CDVPluginResult* pluginResult =  [CDVPluginResult resultWithStatus:CDVCommandStatus_OK];
+                [pluginResult setKeepCallback:[NSNumber numberWithBool:YES]];
+            }];
+        }];
+    }
+}
 
 -(void)onConversionDataReceived:(NSDictionary*) installData {
     
@@ -156,7 +330,7 @@ static NSString *const SUCCESS         = @"Success";
     NSDictionary* errorMessage = @{
                                    @"status": afFailure,
                                    @"type": afOnInstallConversionFailure,
-                                   @"data": _errorMessage
+                                   @"data": _errorMessage.localizedDescription
                                    };
     
     [self performSelectorOnMainThread:@selector(handleCallback:) withObject:errorMessage waitUntilDone:NO];
@@ -179,7 +353,7 @@ static NSString *const SUCCESS         = @"Success";
     NSDictionary* errorMessage = @{
                                    @"status": afFailure,
                                    @"type": afOnAttributionFailure,
-                                   @"data": _errorMessage
+                                   @"data": _errorMessage.localizedDescription
                                    };
     
     [self performSelectorOnMainThread:@selector(handleCallback:) withObject:errorMessage waitUntilDone:NO];
@@ -196,12 +370,13 @@ static NSString *const SUCCESS         = @"Success";
         NSString *jsonMessageStr = [[NSString alloc] initWithBytes:[jsonMessage bytes] length:[jsonMessage length] encoding:NSUTF8StringEncoding];
         
         NSString* status = (NSString*)[message objectForKey: @"status"];
+        NSString* type = (NSString*)[message objectForKey: @"type"];
         
         if([status isEqualToString:afSuccess]){
-            [self reportOnSuccess:jsonMessageStr];
+            [self reportOnSuccess:jsonMessageStr withType:type];
         }
         else{
-            [self reportOnFailure:jsonMessageStr];
+            [self reportOnFailure:jsonMessageStr withType:type];
         }
         
         NSLog(@"jsonMessageStr = %@",jsonMessageStr);
@@ -210,25 +385,53 @@ static NSString *const SUCCESS         = @"Success";
     }
 }
 
--(void) reportOnFailure:(NSString *)errorMessage {
+-(void) reportOnFailure:(NSString *)errorMessage withType:(NSString *)type{
     
-    if(mConversionListener != nil){
-        CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsString:errorMessage];
-        [pluginResult setKeepCallback:[NSNumber numberWithBool:NO]];
-        [self.commandDelegate sendPluginResult:pluginResult callbackId:mConversionListener];
+    if([type isEqualToString:afOnAttributionFailure]){
+        //TODO
+    }
+    else if([type isEqualToString:afOnInstallConversionFailure]){
+        if (mConversionListenerOnResume != nil) {
+            mConversionListenerOnResume = nil;
+        }
         
-        mConversionListener = nil;
-    }}
+        if(mConversionListener != nil){
+            CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsString:errorMessage];
+            [pluginResult setKeepCallback:[NSNumber numberWithBool:NO]];
+            [self.commandDelegate sendPluginResult:pluginResult callbackId:mConversionListener];
+            
+            mConversionListener = nil;
+        }
+    }
+}
 
--(void) reportOnSuccess:(NSString *)data {
+-(void) reportOnSuccess:(NSString *)data withType:(NSString *)type {
     
-    if(mConversionListener != nil){
-        CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsString:data];
-        [pluginResult setKeepCallback:[NSNumber numberWithBool:NO]];
-        [self.commandDelegate sendPluginResult:pluginResult callbackId:mConversionListener];
+    if([type isEqualToString:afOnAppOpenAttribution]){
+        if(mAttributionDataListener != nil){
+            CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsString:data];
+            [pluginResult setKeepCallback:[NSNumber numberWithBool:NO]];
+            [self.commandDelegate sendPluginResult:pluginResult callbackId:mAttributionDataListener];
+            mAttributionDataListener = nil;
+        }
+    }
+    else if([type isEqualToString:afOnInstallConversionDataLoaded]){
+        if (mConversionListenerOnResume != nil) {
+            CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsString:data];
+            [pluginResult setKeepCallback:[NSNumber numberWithBool:NO]];
+            [self.commandDelegate sendPluginResult:pluginResult callbackId:mConversionListenerOnResume];
+            
+            mConversionListenerOnResume = nil;
+        }
         
-        mConversionListener = nil;
-     }
+        if(mConversionListener != nil){
+            CDVPluginResult* pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_OK messageAsString:data];
+            [pluginResult setKeepCallback:[NSNumber numberWithBool:NO]];
+            [self.commandDelegate sendPluginResult:pluginResult callbackId:mConversionListener];
+            
+            mConversionListener = nil;
+        }
+    }
 }
 - (void) handleOpenUrl:(CDVInvokedUrlCommand*)command {
     NSURL *url = [NSURL URLWithString:
@@ -237,5 +440,24 @@ static NSString *const SUCCESS         = @"Success";
     [[AppsFlyerTracker sharedTracker] handleOpenUrl:url options:nil];
 }
 
+
+@end
+
+
+// Universal Links Support - AppDelegate interface:
+@interface AppDelegate (AppsFlyerPlugin)
+
+- (BOOL)application:(UIApplication *)application continueUserActivity:(NSUserActivity *)userActivity restorationHandler:(void (^)(NSArray * _Nullable))restorationHandler;
+
+@end
+
+// Universal Links Support - AppDelegate implementation:
+@implementation AppDelegate (AppsFlyerPlugin)
+
+- (BOOL) application:(UIApplication *)application continueUserActivity:(NSUserActivity *)userActivity restorationHandler:(void (^)(NSArray *_Nullable))restorationHandler
+{
+    [[AppsFlyerTracker sharedTracker] continueUserActivity:userActivity restorationHandler:restorationHandler];
+    return YES;
+}
 
 @end