+++ src/ios/AppsFlyerTracker.h
@@ -2,11 +2,15 @@
 //  AppsFlyerTracker.h
 //  AppsFlyerLib
 //
-//  AppsFlyer iOS SDK 4.7.3
-//  Copyright (c) 2013 AppsFlyer Ltd. All rights reserved.
+//  AppsFlyer iOS SDK 4.8.10 (754)
+//  Copyright (c) 2018 AppsFlyer Ltd. All rights reserved.
 //
 
 #import <Foundation/Foundation.h>
+#import "AppsFlyerCrossPromotionHelper.h"
+#import "AppsFlyerShareInviteHelper.h"
+
+
 
 // In app event names constants
 #define AFEventLevelAchieved            @"af_level_achieved"
@@ -31,12 +35,15 @@
 #define AFEventUpdate                   @"af_update"
 #define AFEventOpenedFromPushNotification @"af_opened_from_push_notification"
 #define AFEventLocation                 @"af_location_coordinates"
-#define AFEventOrderId                  @"af_order_id"
 #define AFEventCustomerSegment          @"af_customer_segment"
 
-
+#define AFEventSubscribe                @"af_subscribe"
+#define AFEventStartTrial               @"af_start_trial"
+#define AFEventAdClick                  @"af_ad_click"
+#define AFEventAdView                   @"af_ad_view"
 
 // In app event parameter names
+#define AFEventParamContent                @"af_content"
 #define AFEventParamAchievenmentId         @"af_achievement_id"
 #define AFEventParamLevel                  @"af_level"
 #define AFEventParamScore                  @"af_score"
@@ -44,7 +51,7 @@
 #define AFEventParamPrice                  @"af_price"
 #define AFEventParamContentType            @"af_content_type"
 #define AFEventParamContentId              @"af_content_id"
-#define AFEventParamContentList            @"ad_content_list"
+#define AFEventParamContentList            @"af_content_list"
 #define AFEventParamCurrency               @"af_currency"
 #define AFEventParamQuantity               @"af_quantity"
 #define AFEventParamRegistrationMethod     @"af_registration_method"
@@ -65,6 +72,7 @@
 #define AFEventParamCustomerUserId         @"af_customer_user_id"
 #define AFEventParamValidated              @"af_validated"
 #define AFEventParamRevenue                @"af_revenue"
+#define AFEventProjectedParamRevenue       @"af_projected_revenue"
 #define AFEventParamReceiptId              @"af_receipt_id"
 #define AFEventParamTutorialId             @"af_tutorial_id"
 #define AFEventParamAchievenmentId         @"af_achievement_id"
@@ -74,6 +82,7 @@
 #define AFEventParamNewVersion             @"af_new_version"
 #define AFEventParamReviewText             @"af_review_text"
 #define AFEventParamCouponCode             @"af_coupon_code"
+#define AFEventParamOrderId                @"af_order_id"
 #define AFEventParam1                      @"af_param_1"
 #define AFEventParam2                      @"af_param_2"
 #define AFEventParam3                      @"af_param_3"
@@ -107,11 +116,21 @@
 #define AFEventParamHotelScore              @"af_hotel_score"
 #define AFEventParamPurchaseCurrency        @"af_purchase_currency"
 
-#define AFEventParamPreferredStarRatings    @"af_preferred_star_ratings"	//array of int (basically a tupple (min,max) but we'll use array of int and instruct the developer to use two values)
+#define AFEventParamPreferredStarRatings    @"af_preferred_star_ratings"    //array of int (basically a tuple (min,max) but we'll use array of int and instruct the developer to use two values)
 
-#define AFEventParamPreferredPriceRange     @"af_preferred_price_range"	//array of int (basically a tupple (min,max) but we'll use array of int and instruct the developer to use two values)
+#define AFEventParamPreferredPriceRange     @"af_preferred_price_range"    //array of int (basically a tuple (min,max) but we'll use array of int and instruct the developer to use two values)
 #define AFEventParamPreferredNeighborhoods  @"af_preferred_neighborhoods" //array of string
 #define AFEventParamPreferredNumStops       @"af_preferred_num_stops"
+
+#define AFEventParamAdRevenueAdType              @"af_adrev_ad_type"
+#define AFEventParamAdRevenueNetworkName         @"af_adrev_network_name"
+#define AFEventParamAdRevenuePlacementId         @"af_adrev_placement_id"
+#define AFEventParamAdRevenueAdSize              @"af_adrev_ad_size"
+#define AFEventParamAdRevenueMediatedNetworkName @"af_adrev_mediated_network_name"
+
+#define kDefaultOneLink          @"go.onelink.me"
+#define kNoOneLinkFallback       @"https://app.appsflyer.com"
+#define kINviteAppleAppID        @"af_siteid"
 
 
 
@@ -136,15 +155,7 @@ typedef enum  {
 
 @end
 
-@interface AppsFlyerTracker : NSObject {
-
-    BOOL _isDebug;
-    BOOL didCollectIAdData;
-    BOOL _useReceiptValidationSandbox;
-    BOOL _useUninstallSandbox;
-    EmailCryptType emailCryptType;
-    NSArray *userEmails;
-}
+@interface AppsFlyerTracker : NSObject
 
 +(AppsFlyerTracker*) sharedTracker;
 
@@ -167,10 +178,6 @@ typedef enum  {
  */
 @property (nonatomic, strong) NSString *currencyCode;
 
-
-/* AppsFlyer's SDK send the data to AppsFlyer's servers over HTTPS. You can set the isHTTPS property to NO in order to use regular HTTP. */
-//@property BOOL isHTTPS;
-
 /* 
  * AppsFLyer SDK collect Apple's advertisingIdentifier if the AdSupport framework included in the SDK.
  * You can disable this behavior by setting the following property to YES.
@@ -184,11 +191,13 @@ typedef enum  {
 @property (nonatomic, setter = setIsDebug:) BOOL isDebug;
 
 
-/*
- * Set this flag to NO, to not collect the device name.
+/*!
+ *  Set this flag to `YES`, to collect the current device name. Default value is `NO`
  */
 @property (nonatomic, setter = setShouldCollectDeviceName:) BOOL shouldCollectDeviceName;
 
+
+@property (nonatomic, setter = setAppInviteOneLink:) NSString* appInviteOneLinkID;
 
 /*
  * Opt-out tracking for specific user
@@ -203,7 +212,7 @@ typedef enum  {
 /*
  * AppsFlyer delegate. See AppsFlyerTrackerDelegate abvoe
  */
-@property (unsafe_unretained, nonatomic) id<AppsFlyerTrackerDelegate> delegate;
+@property (weak, nonatomic) id<AppsFlyerTrackerDelegate> delegate;
 
 /*
  * In app purchase receipt validation Apple environment (production or sandbox). The default value
@@ -278,12 +287,12 @@ typedef enum  {
 /*
  * In case you want to track deep linking, call this method from your delegate's openURL method.
  */
-- (void) handleOpenURL:(NSURL *)url sourceApplication:(NSString *)sourceApplication __attribute__((deprecated));
+- (void) handleOpenURL:(NSURL *)url sourceApplication:(NSString *)sourceApplication;
 
 /*
  * In case you want to track deep linking, call this method from your delegate's openURL method with refferer.
  */
-- (void) handleOpenURL:(NSURL *)url sourceApplication:(NSString *)sourceApplication withAnnotation:(id) annotation __attribute__((deprecated));
+- (void) handleOpenURL:(NSURL *)url sourceApplication:(NSString *)sourceApplication withAnnotation:(id) annotation;
 
 
 - (void) handleOpenUrl:(NSURL *) url options:(NSDictionary *)options;
@@ -309,5 +318,29 @@ typedef enum  {
 
 
 - (void) remoteDebuggingCallWithData:(NSString *) data;
+
+/*!
+ *  @brief This property accepts a string value representing the host name for all enpoints.
+ *  @warning To use `default` SDK endpoint – set value to `nil`.
+ *  @code
+ *  Objective-C:
+ *  [[AppsFlyerTracker sharedTracker] setHost:@"example.com"];
+ *  Swift:
+ *  AppsFlyerTracker.shared().host = "example.com"
+ *  @endcode
+ */
+
+@property (nonatomic, strong) NSString *host;
+
+/*!
+ *  This property is responsible for timeout between sessions in seconds.
+ *  Default value is 5 seconds.
+ */
+@property (atomic) NSUInteger minTimeBetweenSessions;
+
+/*!
+ *  WARNING! This will disable all requests from AppsFlyer SDK
+ */
+@property (atomic) BOOL isStopTracking;
 
 @end