+++ RELEASENOTES.md
@@ -1,4 +1,81 @@
 # Release Notes
+
+### 4.4.12
+Release date **Nov 14 2018**
+Release type: Major / Minor / **Hotfix**
+
+**Overview and Highlights:**
+
+- Fixed Android platform <source-file /> tag to work with Cordova-Android 7.1.0 and above
+
+- Updated Android SDK to v4.8.18
+
+- Updated iOS SDK to v4.8.10
+
+
+### 4.4.10
+Release date: **Nov 01 2018**
+Release type: Major / **Minor** / Hotfix
+
+**Overview and Highlights:**
+
+- Added new method API for iOS: `registerUninstall`
+
+- Updated Android SDK to latest 4.8.17
+
+- Updated iOS SDK to latest 4.8.10
+
+
+---
+
+
+### 4.4.8
+Release date: May 10 2018
+
+Release type: Major / Minor / **Hotfix**
+
+Overview and Highlights
+
+- [Android: invalid JSON in initSdk's onSuccess callback on link click](https://github.com/AppsFlyerSDK/cordova-plugin-appsflyer-sdk/issues/31)
+
+
+
+---
+
+
+
+
+### 4.4.7
+Release date: May 01 2018
+
+Release type: Major / **Minor** / Hotfix
+
+Overview and Highlights
+
+- Android: updated  to latest SDK: 4.8.10
+
+- Android: Moved from jar to gradle 
+
+- Android: added Google install referrer
+
+- Android: Added 2 APIs: collectAndroidID and collectIMEI
+
+- stopTracking API
+
+
+
+---
+
+
+
+
+### 4.4.0
+* changed Android paths to support cordova-android 7.0.0
+
+### 4.3.0
+* Added User Invite and Cross Promotion API Calls
+* Bug Fixes
+
 ### 4.2.10 (Nov 30, 2016)
 * updated iOS Appsflyer SDK vertsion to 4.5.12
 * updated Android Appsflyer SDK vertsion to 4.6.1